Configuration for forked base chain.
This initial commit creates the Docker configurations required to run
a forked basechain node with Foundry's "anvil" tool.  We also set up
an nginx reverse proxy so that we can add the required
"Content-Type: application/json" header always, and which we will use
in the future to switch between the basechain RPC and the utility RPC server.